fixed animation issues
