+++ feature/dogdetails/src/main/java/com/sidgowda/pawcalc/dogdetails/ui/HumanYearsChart.kt
@@ -73,7 +73,6 @@ internal fun HumanYearsChart(
         val convertSweepAngle = convertedSweepAngle(angle.value).toDouble()
         val x = convertedRadiusX(radius, convertSweepAngle) + 5.dp.toPx()
         val y = convertedRadiusY(radius, convertSweepAngle) + 5.dp.toPx()
-
         translate(
             left = x,
             top = y
@@ -84,44 +83,6 @@ internal fun HumanYearsChart(
                 )
             }
         }
-    }
-}
-
-fun convertedSweepAngle(sweepAngle: Float): Float {
-    return when {
-        sweepAngle >= 0f && sweepAngle <= 90f -> 90 - sweepAngle
-        sweepAngle >= 91f && sweepAngle <= 180f -> 360 - (sweepAngle - 90)
-        sweepAngle >= 181f && sweepAngle <= 270f -> 270 - (sweepAngle - 180)
-        sweepAngle >= 271f && sweepAngle <= 360f -> 180 - (sweepAngle - 270)
-        else -> 0f
-    }
-}
-fun convertedRadiusX(radius: Float, sweepAngle: Double): Float {
-   return if (sweepAngle >= 0f && sweepAngle <= 90f) {
-        radius * kotlin.math.cos(Math.toRadians(sweepAngle)).toFloat() + radius
-    } else if (sweepAngle > 90f && sweepAngle <= 180f) {
-       radius + radius * kotlin.math.cos(Math.toRadians(sweepAngle)).toFloat()
-    } else if (sweepAngle > 180f && sweepAngle <= 270f) {
-      radius - radius * -kotlin.math.cos(Math.toRadians(sweepAngle)).toFloat()
-   } else if (sweepAngle > 270f && sweepAngle <= 360f) {
-       radius * kotlin.math.cos(Math.toRadians(sweepAngle)).toFloat() + radius
-   }
-   else {
-        0f
-   }
-}
-
-fun convertedRadiusY(radius: Float, sweepAngle: Double): Float {
-    return if (sweepAngle >= 0f && sweepAngle <= 90f) {
-        radius - radius * kotlin.math.sin(Math.toRadians(sweepAngle)).toFloat()
-    } else if (sweepAngle > 90f && sweepAngle <= 180f) {
-        radius + radius * - kotlin.math.sin(Math.toRadians(sweepAngle)).toFloat()
-    } else if (sweepAngle > 180f && sweepAngle <= 270f) {
-        radius + -radius * kotlin.math.sin(Math.toRadians(sweepAngle)).toFloat()
-    } else if (sweepAngle > 270f && sweepAngle <= 360f) {
-        radius + -radius * kotlin.math.sin(Math.toRadians(sweepAngle)).toFloat()
-    } else {
-        0f
     }
 }
